# docs/swagger.yaml
definitions:
  handlers.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  handlers.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handlers.StatusResponse:
    properties:
      status:
        type: string
    type: object
  handlers.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      gender:
        type: string
      name:
        type: string
    type: object
  models.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      description: Returns a basic message confirming the API is reachable.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Root message
      tags:
      - Health
  /api/users:
    post:
      consumes:
      - application/json
      description: Creates a new account and returns a JWT token for the user.
      parameters:
      - description: User sign up payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/users/profile:
    get:
      description: Returns the authenticated user's profile details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Applies partial or full updates to the authenticated user's profile.
      parameters:
      - description: Profile update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
  /api/users/profile/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an avatar image for the current user and updates the profile.
      parameters:
      - description: Avatar image
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload avatar image
      tags:
      - Users
  /api/users/sign_in:
    post:
      consumes:
      - application/json
      description: Verifies credentials and returns a JWT token.
      parameters:
      - description: User sign in payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Authenticate a user
      tags:
      - Auth
  /api/users/sign_out:
    delete:
      description: Stateless logout helper that simply acknowledges the request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      security:
      - BearerAuth: []
      summary: Sign out the current user
      tags:
      - Auth
  /health:
    get:
      description: Returns OK when the service is healthy.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Health check
      tags:
      - Health
swagger: "2.0"